Replace the default indexer for ib_counter_t values for performance
Summary: The default indexer for ib_counter_t values uses the result from 'pthread_self() % 64' to index into an array.  Since pthread_self() likely returns a pointer to an internal structure on 64-bit platforms 'pthread_self() % 64' will often (always?) result in 0.  This diff switches to using a different indexer that uses the current CPU when the sched_getcpu() function is available.

Test Plan: MTR

Reviewers: yoshinorim, MarkCallaghan

Reviewed By: MarkCallaghan

Subscribers: webscalesql-eng

Differential Revision: https://reviews.facebook.net/D65427

+++ include/ut0counter.h
@@ -55,7 +55,8 @@ struct generic_indexer_t {
 };
 
 #ifdef HAVE_SCHED_GETCPU
-#include <utmpx.h>
+//#include <utmpx.h>  // Including this causes problems with EMPTY symbol
+#include <sched.h>  // Include this instead
 /** Use the cpu id to index into the counter array. If it fails then
 use the thread id. */
 template <typename Type, int N>
@@ -66,7 +67,7 @@ struct get_sched_indexer_t : public generic_indexer_t<Type, N> {
 	size_t get_rnd_index() const {
 
 		size_t	cpu = sched_getcpu();
-		if (cpu == -1) {
+		if (cpu == (size_t) -1) {
 			cpu = get_curr_thread_id();
 		}
